requirements.txt
streamlit
pandas
numpy
base64
path
datetime
os
io
openpyxl
python-docx
python-magic
